requirements.txt
openpyxl
requests
fitz
pytesseract
Pillow
word2number
transformers
torch
torchvision